Queue a job to be run next.

+++ app/controllers/delayed/web/jobs_controller.rb
@@ -1,9 +1,24 @@
 module Delayed
   module Web
     class JobsController < Delayed::Web::ApplicationController
+      def queue
+        if job.can_queue?
+          job.queue!
+          flash[:notice] = t(:notice, scope: 'delayed/web.flashes.jobs.queued')
+        else
+          flash[:alert] = t(:alert, scope: 'delayed/web.flashes.jobs.queued', status: job.status)
+        end
+        redirect_to jobs_path
+      end
+
       def destroy
-        job.destroy
-        redirect_to jobs_path, notice: t(:notice, scope: 'delayed/web.flashes.jobs.destroyed')
+        if job.can_destroy?
+          job.destroy
+          flash[:notice] = t(:notice, scope: 'delayed/web.flashes.jobs.destroyed')
+        else
+          flash[:alert] = t(:alert, scope: 'delayed/web.flashes.jobs.destroyed', status: job.status)
+        end
+        redirect_to jobs_path
       end
 
     private
